feat: move instance types apply unit tests to correct package
remove apply related unit tests from general instancetype_test.go
file to apply package.
