+++ pkg/instancetype/apply/apply_suite_test.go
@@ -1,0 +1,13 @@
+package apply_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestApply(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Apply Suite")
+}
