--- pkg/instancetype/instancetype_suite_test.go
@@ -1,11 +1,0 @@
-package instancetype_test
-
-import (
-	"testing"
-
-	"kubevirt.io/client-go/testutils"
-)
-
-func TestInstancetype(t *testing.T) {
-	testutils.KubeVirtTestSuiteSetup(t)
-}
