+++ pkg/instancetype/preference/apply/machine_test.go
@@ -1,0 +1,41 @@
+package apply_test
+
+import (
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+
+	k8sfield "k8s.io/apimachinery/pkg/util/validation/field"
+
+	virtv1 "kubevirt.io/api/core/v1"
+	v1beta1 "kubevirt.io/api/instancetype/v1beta1"
+
+	"kubevirt.io/kubevirt/pkg/instancetype/apply"
+	"kubevirt.io/kubevirt/pkg/libvmi"
+)
+
+var _ = Describe("Preference - Apply to vmi - machine", func() {
+	var (
+		vmi              *virtv1.VirtualMachineInstance
+		instancetypeSpec *v1beta1.VirtualMachineInstancetypeSpec
+		preferenceSpec   *v1beta1.VirtualMachinePreferenceSpec
+
+		field      = k8sfield.NewPath("spec", "template", "spec")
+		vmiApplier = apply.NewVMIApplier()
+	)
+
+	BeforeEach(func() {
+		vmi = libvmi.New()
+	})
+
+	It("should apply to VMI", func() {
+		preferenceSpec = &v1beta1.VirtualMachinePreferenceSpec{
+			Machine: &v1beta1.MachinePreferences{
+				PreferredMachineType: "q35-rhel-8.0",
+			},
+		}
+
+		Expect(vmiApplier.ApplyToVMI(field, instancetypeSpec, preferenceSpec, &vmi.Spec, &vmi.ObjectMeta)).To(Succeed())
+
+		Expect(vmi.Spec.Domain.Machine.Type).To(Equal(preferenceSpec.Machine.PreferredMachineType))
+	})
+})
